Adds container wrapper with iterator interface
In many cases BWAPI returns whole containers instead of
references or iterators to them.

For example, UnitIterface::getLoadedUnits() returns Unitset.

We take ownership over the container object and then
provide a generic iterator interface to it.

Because of move semantics, container internals are moved, not copied.

Also we provide wrapper function and some template magic to deduce
mapping between iterator template specialization and it's API counterpart.

This helps to avoid interface misuse and makes it more friendly.